chore: 파일 위치 변경 및 코드 일관화 정리
