+++ src/app/album/[id]/_component/PhotoAddButton.tsx
@@ -1,5 +1,6 @@
 "use client"
-import { Icon } from "@/components"
+
+import Icon from "@/common/Icon"
 
 export const PhotoAddButton = () => {
   const onClick = () => {
