+++ src/app/api/myfetch/customFetch.ts
