+++ src/app/scanner/page.tsx
@@ -2,7 +2,7 @@
 
 import { IDetectedBarcode, Scanner } from "@yudiel/react-qr-scanner"
 
-import { isUrlIncluded } from "@/libs/qr"
+import { isUrlIncluded } from "@/libs"
 
 import { postQrCode } from "../api/photo"
 
@@ -20,12 +20,15 @@ const style = {
   },
 }
 
-const QrPage = () => {
+const ScannerPage = () => {
   const onScan = async (result: IDetectedBarcode[]) => {
     const { rawValue } = result[0]
 
     if (!isUrlIncluded(rawValue)) {
-      alert("지원하지 않는 QR코드입니다.")
+      if (confirm("지원하지 않는 QR코드입니다. 웹사이트를 열어드릴까요?")) {
+        window.open(rawValue, "_blank")
+      }
+
       return
     }
 
@@ -73,4 +76,4 @@ const QrPage = () => {
   )
 }
 
-export default QrPage
+export default ScannerPage
