+++ src/common/AlbumItem.tsx
@@ -1,53 +1,20 @@
 "use client"
 
-import { cva, VariantProps } from "class-variance-authority"
+import { VariantProps } from "class-variance-authority"
 import { ChangeEvent, useState } from "react"
 
-import { Badge, ColorIcon } from "@/components"
-import type { albumValue } from "@/types"
+import { AlbumValue } from "@/app/album/types"
+import { ALBUM_ITEM_STYLES, PHOTO_COUNT_STYLES } from "@/constants/styles"
 import { cn } from "@/utils"
 
-const albumItemVariants = cva("w-[164px] h-[150px] rounded-2xl relative p-4", {
-  variants: {
-    type: {
-      HEART: "bg-red-200",
-      FIRE: "bg-butter-200",
-      BASKETBALL: "bg-green-200",
-      BUILDING: "bg-blue-200",
-      STARFALL: "bg-purple-200",
-      SMILE_FACE: "bg-pink-200",
-    },
-
-    isEditable: {
-      true: "w-60 h-[219.51px] rounded-[23.41px]",
-    },
-  },
-  defaultVariants: {
-    type: "HEART",
-  },
-})
-
-const photoCountVariants = cva("text-caption-1 font-medium", {
-  variants: {
-    type: {
-      HEART: "text-red-500",
-      FIRE: "text-orange-600",
-      BASKETBALL: "text-green-700",
-      BUILDING: "text-sky-blue-700",
-      STARFALL: "text-purple-600",
-      SMILE_FACE: "text-pink-600",
-    },
-  },
-  defaultVariants: {
-    type: "HEART",
-  },
-})
+import Badge from "./Badge"
+import ColorIcon from "./ColorIcon"
 
 export interface AlbumItemProps
   extends React.HTMLAttributes<HTMLDivElement>,
-    VariantProps<typeof albumItemVariants> {
-  value: albumValue
-  handleValue: (v: albumValue) => void
+    VariantProps<typeof ALBUM_ITEM_STYLES> {
+  value: AlbumValue
+  handleValue: (v: AlbumValue) => void
 }
 
 /**
@@ -56,7 +23,7 @@ export interface AlbumItemProps
  * @param handleValue album 내부에서 사용되는 값을 handle할 수 있는 값
  * @param className className을 추가하여 추가적인 스타일링도 가능합니다.
  */
-function AlbumItem({ value, handleValue, ...props }: AlbumItemProps) {
+const AlbumItem = ({ value, handleValue, ...props }: AlbumItemProps) => {
   const {
     name: initialName,
     type,
@@ -82,7 +49,7 @@ function AlbumItem({ value, handleValue, ...props }: AlbumItemProps) {
 
   return (
     <div
-      className={`${cn(albumItemVariants({ type, isEditable }), className)}`}
+      className={`${cn(ALBUM_ITEM_STYLES({ type, isEditable }), className)}`}
       {...props}>
       {isEditable ? (
         <>
@@ -100,7 +67,7 @@ function AlbumItem({ value, handleValue, ...props }: AlbumItemProps) {
       ) : (
         <>
           <h3 className="text-title-2 font-bold text-gray-800">{name}</h3>
-          <div className={`${cn(photoCountVariants({ type }))}`}>
+          <div className={`${cn(PHOTO_COUNT_STYLES({ type }))}`}>
             사진 {photoCount}장
           </div>
 
@@ -121,4 +88,4 @@ function AlbumItem({ value, handleValue, ...props }: AlbumItemProps) {
   )
 }
 
-export { AlbumItem }
+export default AlbumItem
