+++ src/common/FloatingButton.tsx
@@ -1,0 +1,7 @@
+import Button, { ButtonProps as FloatingButtonProps } from "./Button"
+
+const FloatingButton = ({ className, ...props }: FloatingButtonProps) => {
+  return <Button className={`${className} rounded-full`} {...props} />
+}
+
+export default FloatingButton
