+++ src/common/QueryProviders.tsx
@@ -38,11 +38,7 @@ function getQueryClient() {
   }
 }
 
-export default function QueryProviders({
-  children,
-}: {
-  children: React.ReactNode
-}) {
+export function QueryProviders({ children }: { children: React.ReactNode }) {
   // NOTE: Avoid useState when initializing the query client if you don't
   //       have a suspense boundary between this and the code that may
   //       suspend because React will throw away the client on the initial
