--- src/common/Test.tsx
@@ -1,16 +1,0 @@
-"use client"
-
-import { useSuspenseQuery } from "@tanstack/react-query"
-
-import { getJsonplaceholder } from "@/app/api"
-
-const Test = () => {
-  const { data } = useSuspenseQuery({
-    queryKey: ["test", "asd"],
-    queryFn: () => getJsonplaceholder("1"),
-  })
-
-  return <div>get list....{data?.title}</div>
-}
-
-export default Test
