--- src/components/FloatingButton.tsx
@@ -1,9 +1,0 @@
-import { Button } from "@/components"
-
-import { ButtonProps as FloatingButtonProps } from "./Button"
-
-function FloatingButton({ className, ...props }: FloatingButtonProps) {
-  return <Button className={`${className} rounded-full`} {...props} />
-}
-
-export { FloatingButton }
