--- src/libs/cookie.ts
@@ -1,3 +1,0 @@
-import Cookies from "js-cookie"
-
-export const getAccessToken = () => Cookies.get("accessToken")
